defaillances.js --> TS
