+++ dbmongo/js/reduce.algo2/ava_tests.ts
@@ -11,7 +11,7 @@ import { flatten } from "./flatten"
 import { outputs } from "./outputs"
 import { repeatable } from "./repeatable"
 import { add } from "./add"
-import { defaillances } from "./defaillances.js"
+import { defaillances } from "./defaillances"
 import { dealWithProcols } from "./dealWithProcols.js"
 import { populateNafAndApe } from "./populateNafAndApe"
 import { cotisation } from "./cotisation.js"
