--- dbmongo/js/reduce.algo2/defaillances.d.ts
@@ -1,4 +1,0 @@
-export function defaillances(
-  v: { altares: any; procol: any },
-  output_indexed: {}
-): void
