--- dbmongo/js/reduce.algo2/defaillances.js
@@ -1,7 +1,0 @@
-function defaillances(v, output_indexed) {
-  "use strict"
-  f.dealWithProcols(v.altares, "altares", output_indexed)
-  f.dealWithProcols(v.procol, "procol", output_indexed)
-}
-
-exports.defaillances = defaillances
